Fixed fort multiplier. Still debugging building costs reduction.
